+++ GLideN64/projects/msvc12/GLideN64.vcxproj.filters
@@ -98,34 +98,34 @@
     <ClCompile Include="..\..\src\VI.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\F3D.cpp">
+    <ClCompile Include="..\..\src\uCodes\F3D.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\F3DDKR.cpp">
+    <ClCompile Include="..\..\src\uCodes\F3DDKR.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\F3DEX.cpp">
+    <ClCompile Include="..\..\src\uCodes\F3DEX.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\F3DEX2.cpp">
+    <ClCompile Include="..\..\src\uCodes\F3DEX2.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\F3DPD.cpp">
+    <ClCompile Include="..\..\src\uCodes\F3DPD.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\L3D.cpp">
+    <ClCompile Include="..\..\src\uCodes\L3D.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\L3DEX.cpp">
+    <ClCompile Include="..\..\src\uCodes\L3DEX.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\L3DEX2.cpp">
+    <ClCompile Include="..\..\src\uCodes\L3DEX2.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\S2DEX.cpp">
+    <ClCompile Include="..\..\src\uCodes\S2DEX.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\S2DEX2.cpp">
+    <ClCompile Include="..\..\src\uCodes\S2DEX2.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\3DMath.cpp">
@@ -164,7 +164,7 @@
     <ClCompile Include="..\..\src\windows\ZilmarAPIImpl_windows.cpp">
       <Filter>Source Files\windows</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\F3DEX2CBFD.cpp">
+    <ClCompile Include="..\..\src\uCodes\F3DEX2CBFD.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\windows\Config_windows.cpp">
@@ -173,10 +173,10 @@
     <ClCompile Include="..\..\src\mupenplus\Config_mupenplus.cpp">
       <Filter>Source Files\mupenplus</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\Turbo3D.cpp">
+    <ClCompile Include="..\..\src\uCodes\Turbo3D.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\ZSort.cpp">
+    <ClCompile Include="..\..\src\uCodes\ZSort.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\TextDrawer.cpp">
@@ -254,16 +254,16 @@
     <ClCompile Include="..\..\src\BufferCopy\ColorBufferToRDRAM_BufferStorageExt.cpp">
       <Filter>Source Files\BufferCopy</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\F3DSETA.cpp">
+    <ClCompile Include="..\..\src\uCodes\F3DSETA.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\F3DGOLDEN.cpp">
+    <ClCompile Include="..\..\src\uCodes\F3DGOLDEN.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\F3DEX2MM.cpp">
+    <ClCompile Include="..\..\src\uCodes\F3DZEX2.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\F3DBETA.cpp">
+    <ClCompile Include="..\..\src\uCodes\F3DBETA.cpp">
       <Filter>Source Files\uCodes</Filter>
     </ClCompile>
   </ItemGroup>
@@ -325,34 +325,34 @@
     <ClInclude Include="..\..\src\VI.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\F3D.h">
+    <ClInclude Include="..\..\src\uCodes\F3D.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\F3DDKR.h">
+    <ClInclude Include="..\..\src\uCodes\F3DDKR.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\F3DEX.h">
+    <ClInclude Include="..\..\src\uCodes\F3DEX.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\F3DEX2.h">
+    <ClInclude Include="..\..\src\uCodes\F3DEX2.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\F3DPD.h">
+    <ClInclude Include="..\..\src\uCodes\F3DPD.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\L3D.h">
+    <ClInclude Include="..\..\src\uCodes\L3D.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\L3DEX.h">
+    <ClInclude Include="..\..\src\uCodes\L3DEX.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\L3DEX2.h">
+    <ClInclude Include="..\..\src\uCodes\L3DEX2.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\S2DEX.h">
+    <ClInclude Include="..\..\src\uCodes\S2DEX.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\S2DEX2.h">
+    <ClInclude Include="..\..\src\uCodes\S2DEX2.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\ZilmarGFX_1_3.h">
@@ -370,13 +370,13 @@
     <ClInclude Include="..\..\src\Log.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\F3DEX2CBFD.h">
+    <ClInclude Include="..\..\src\uCodes\F3DEX2CBFD.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\Turbo3D.h">
+    <ClInclude Include="..\..\src\uCodes\Turbo3D.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\ZSort.h">
+    <ClInclude Include="..\..\src\uCodes\ZSort.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\GLideNHQ\Ext_TxFilter.h">
@@ -463,16 +463,16 @@
     <ClInclude Include="..\..\src\BufferCopy\ColorBufferToRDRAM_GL.h">
       <Filter>Header Files\BufferCopy</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\F3DSETA.h">
+    <ClInclude Include="..\..\src\uCodes\F3DSETA.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\F3DGOLDEN.h">
+    <ClInclude Include="..\..\src\uCodes\F3DGOLDEN.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\F3DEX2MM.h">
+    <ClInclude Include="..\..\src\uCodes\F3DZEX2.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\F3DBETA.h">
+    <ClInclude Include="..\..\src\uCodes\F3DBETA.h">
       <Filter>Header Files\uCodes</Filter>
     </ClInclude>
   </ItemGroup>
