--- GLideN64/src/F3DEX2MM.h
@@ -1,6 +1,0 @@
-#ifndef F3DEX2MM_H
-#define F3DEX2MM_H
-
-void F3DEX2MM_Init();
-
-#endif // F3DEX2M_H
