+++ GLideN64/src/RSP.cpp
@@ -3,8 +3,8 @@
 #include "RSP.h"
 #include "RDP.h"
 #include "N64.h"
-#include "F3D.h"
-#include "Turbo3D.h"
+#include "uCodes/F3D.h"
+#include "uCodes/Turbo3D.h"
 #include "VI.h"
 #include "Combiner.h"
 #include "FrameBuffer.h"
