+++ GLideN64/src/uCodes/F3DZEX2.h
@@ -1,0 +1,6 @@
+#ifndef F3DZEX2_H
+#define F3DZEX2_H
+
+void F3DZEX2_Init();
+
+#endif // F3DEX2M_H
