+++ Makefile.common
@@ -240,16 +240,16 @@ SOURCES_CXX   += \
 	$(VIDEODIR_GLIDEN64)/src/Config.cpp \
 	$(VIDEODIR_GLIDEN64)/src/convert.cpp \
 	$(VIDEODIR_GLIDEN64)/src/DepthBuffer.cpp \
-	$(VIDEODIR_GLIDEN64)/src/F3D.cpp \
-	$(VIDEODIR_GLIDEN64)/src/F3DDKR.cpp \
-	$(VIDEODIR_GLIDEN64)/src/F3DEX.cpp \
-	$(VIDEODIR_GLIDEN64)/src/F3DPD.cpp \
-	$(VIDEODIR_GLIDEN64)/src/F3DGOLDEN.cpp \
-	$(VIDEODIR_GLIDEN64)/src/F3DBETA.cpp \
-	$(VIDEODIR_GLIDEN64)/src/F3DSETA.cpp \
-	$(VIDEODIR_GLIDEN64)/src/F3DEX2.cpp \
-	$(VIDEODIR_GLIDEN64)/src/F3DEX2MM.cpp \
-	$(VIDEODIR_GLIDEN64)/src/F3DEX2CBFD.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/F3D.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/F3DDKR.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/F3DEX.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/F3DPD.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/F3DGOLDEN.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/F3DBETA.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/F3DSETA.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/F3DEX2.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/F3DZEX2.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/F3DEX2CBFD.cpp \
 	$(VIDEODIR_GLIDEN64)/src/FBOTextureFormats.cpp \
 	$(VIDEODIR_GLIDEN64)/src/FrameBuffer.cpp \
 	$(VIDEODIR_GLIDEN64)/src/FrameBufferInfo.cpp \
@@ -258,18 +258,18 @@ SOURCES_CXX   += \
 	$(VIDEODIR_GLIDEN64)/src/GLideN64.cpp \
 	$(VIDEODIR_GLIDEN64)/src/gSP.cpp \
 	$(VIDEODIR_GLIDEN64)/src/Keys.cpp \
-	$(VIDEODIR_GLIDEN64)/src/L3D.cpp \
-	$(VIDEODIR_GLIDEN64)/src/L3DEX2.cpp \
-	$(VIDEODIR_GLIDEN64)/src/L3DEX.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/L3D.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/L3DEX2.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/L3DEX.cpp \
 	$(VIDEODIR_GLIDEN64)/src/N64.cpp \
 	$(VIDEODIR_GLIDEN64)/src/OpenGL.cpp \
 	$(VIDEODIR_GLIDEN64)/src/Performance.cpp \
 	$(VIDEODIR_GLIDEN64)/src/RDP.cpp \
 	$(VIDEODIR_GLIDEN64)/src/RSP.cpp \
-	$(VIDEODIR_GLIDEN64)/src/S2DEX2.cpp \
-	$(VIDEODIR_GLIDEN64)/src/S2DEX.cpp \
-	$(VIDEODIR_GLIDEN64)/src/Turbo3D.cpp \
-	$(VIDEODIR_GLIDEN64)/src/ZSort.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/S2DEX2.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/S2DEX.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/Turbo3D.cpp \
+	$(VIDEODIR_GLIDEN64)/src/uCodes/ZSort.cpp \
 	$(VIDEODIR_GLIDEN64)/src/ShaderUtils.cpp \
 	$(VIDEODIR_GLIDEN64)/src/SoftwareRender.cpp \
 	$(VIDEODIR_GLIDEN64)/src/Textures.cpp \
